==> 01 Merge Coordenadas e Dados Primeira Infância .ipynb <==
import pandas as pd

coordenadas = pd.read_csv('dados/TODAS_COORDENADAS.txt', sep = ';',header= 0 )

coordenadas.drop('index', axis=1,inplace = True)

coordenadas.columns = ['ID_FAMILIA_CADUNICO', 'LAT', 'LOG']

data = pd.read_csv('dados/1a_Infancia_familias_cadunico_v2G.csv', encoding='iso-8859-1', sep = ';')

cod_ibge = pd.read_csv('dados/Municipios-Brasileiros-master/csv/municipios.csv')

#data = pd.read_csv('dados/1a_infancia.csv', encoding='iso-8859-1', sep = ';')

new_df = coordenadas.merge(data, on = 'ID_FAMILIA_CADUNICO')

new_df = cod_ibge.merge(new_df, left_on = 'codigo_ibge',right_on ='CODG_IBGE_MUN_RESIDE')

dados = new_df.to_csv('dados/1a_infancia_300k.csv',index = False)
