==> 02 Tratando dados Primeira Infância completo 300K.ipynb <==
import pandas as pd
import numpy as np
from unicodedata import normalize

dados = pd.read_csv('dados/1a_infancia_300k.csv')

dados.head()

dados.drop(columns = ['codigo_ibge', 'capital','codigo_uf','NOME_UNIDADE_TERRITAL'], inplace = True)

#goias = pd.read_csv('dados/latitude_longitude_municipos_goianos.csv')

### Tratando os Dados

#### Normalizando nome cidade

goias['nome'] =  goias['nome'].str.upper() #Transformando nome da cidade em maiusculo

nome_cidade = []
for i in goias['nome']:
    source = i
    target = normalize('NFKD', source).encode('ASCII','ignore').decode('ASCII')
    nome_cidade.append(target)
    
goias['nome_tratado'] = nome_cidade  # Retirando acentuação

goias.head()

### Exportando dados

novo_primeira_infancia_tratado = dados.to_csv ('dados/primeira_infancia_300k_tratado.csv', index = None, header=True)
novo_primeira_infancia_tratado
