==> 03 comparando coordenadas do endereço x coordenadas cidade 300k.ipynb <==
import  pandas as pd
import time, os
from random import randint
import geopy.distance
import geopy
from geopy.geocoders import Nominatim

# Comparando coordenadas endereço com coordenadas cidade usando geopy

dados = pd.read_csv('dados/primeira_infancia_300k_tratado.csv')

dados

df1 = pd.DataFrame(columns=['ID_FAMILIA_CADUNICO','DESVIO'])
retornos = 0
retate1km = 0
for index, row in dados.iterrows():
    try:
        
        if row["LAT"] != "None" and row["LAT"] != "ERROR" and row["latitude"] != "None" and row["latitude"] != "ERROR":
            retornos=retornos + 1
            
            p1 = (float(row["LAT"]),float(row["LOG"]))
            p2 = (float(row["latitude"]),float(row["longitude"]))

            distkm = geopy.distance.distance(p1, p2).km
            
            df1.loc[index] = [row['ID_FAMILIA_CADUNICO'], distkm]
            
            if distkm <= 1:
                retate1km = retate1km + 1
            #print("%.3f" %(geopy.distance.distance(p1, p2).km))
    except:
    #else:
        df1.loc[index] = [row['ID_FAMILIA_CADUNICO'],"Nan"]
        #print(None)


df1.describe().round(2)

novo_dados = pd.concat([dados, df1['DESVIO']], axis=1)

novo_dados.isnull().sum()

## Alterando os valores com erro por Lat, Lon da Cidade

novo_dados['aproximação'] = novo_dados['DESVIO'].apply(lambda x: 'Erro' if float(x) >= 10.0 else 'ok')

estatistica = novo_dados['aproximação'] == 'Erro'
porcentagem_erro = estatistica.sum() / len(novo_dados)*100
print('Valores fora do perimetro: '+ str(porcentagem_erro.round(2)) +'%')

dados_desvio = novo_dados.to_csv('dados/dados_primeira_infancia_300k_com_desvio.csv', index = None, header=True)

### Filtros 

def filrando_por_cidade(nome_cidade:'recebe o nome do municipo'):
    selecao = (novo_dados['Municípios.NOME_MUNICIPIO'] == nome_cidade)
    df1 = novo_dados[selecao]
    d = df1['aproximação'].value_counts()
    return d

filrando_por_cidade('ALEXANIA')

filrando_por_cidade('ABADIA DE GOIAS')

selecao = (novo_dados['Municípios.NOME_MUNICIPIO'] == 'ALEXANIA')
df1 = novo_dados[selecao]

df1['aproximação'].value_counts()

filrando_por_cidade('ALEXANIA')

import seaborn as sns
df = novo_dados['lat_cidade']
x =  novo_dados['DESVIO']
 
sns.boxplot(data=df, saturation=1, orient = 'h')
#sns.plt.show()
