==> 04 Substitui valores com erro .ipynb <==
import  pandas as pd
import plotly.graph_objects as go
import random

dados = pd.read_csv('dados/dados_primeira_infancia_300k_com_desvio.csv')

dados.head(1)

### solução para gerar coordenada dinâmica

#num_randon = []
#for d, i in dados.iterrows():
#    if i['aproximação'] == 'Erro':
#        num_randon.append(random.uniform(-0.001, 0.0100))
#    else:
#        num_randon.append(0)

selecao = (dados['aproximação'] == 'Erro')
df1 = dados[selecao]

lat_radom = []
for d, i in df1.iterrows():
    lat_radom.append(random.uniform(-0.001, 0.0100))
    

lon_radom = []
for d, i in df1.iterrows():
    lon_radom.append(random.uniform(0.001, -0.0100))

df1['lat_corrigido'] = df1['latitude'] + lat_radom
df1['lon_corrigido'] = df1['longitude'] + lon_radom

selecao = (dados['aproximação'] == 'ok')
df2 = dados[selecao]

df2['lat_corrigido'] = dados['LAT']    #mantendo os dados corretos de lat e long
df2['lon_corrigido'] = dados['LOG']

novo_dados = pd.concat([df1,df2], ignore_index=True)  #juntando os dados
novo_dados

#### Novo arquivo CSV tratado

novo_dados.dropna(subset=['lat_corrigido', 'lon_corrigido'],inplace = True)

novo_dados.info()

novo_dados = novo_dados.to_csv('dados/novo_dados_primeira_300k_infancia.csv')
novo_dados

## Gráficos

### Folium

import folium

novo_dados.dropna(subset=['lat_corrigido','lon_corrigido'],inplace = True)

Goias = folium.Map(
    location=[-16.0908158,-48.5083711],
    zoom_start=7
)
for _, loc in novo_dados.iterrows():
    folium.Marker(
        location=[loc['lat_corrigido'], loc['lon_corrigido']],
        
    ).add_to(Goias)

#Goias

NOME_MUNICIPIO = novo_dados['NOME_MUNICIPIO'] == 'ABADIA DE GOIAS'
ABADIA_DE_GOIAS = novo_dados[NOME_MUNICIPIO]

ABADIA_DE_GOIAS

site_lat = ABADIA_DE_GOIAS['lat_corrigido']
site_lon = ABADIA_DE_GOIAS['lon_corrigido']


fig = go.Figure(go.Scattermapbox(
    
        lat=site_lat,
        lon=site_lon,
        mode='markers',
        marker=go.scattermapbox.Marker(
            size=9
        ),
        text = ABADIA_DE_GOIAS['ENDERECO'],
        #text = dados['DESC_CONDICAO_CADASTRO'],
    
    
    ))

fig.update_layout(
    autosize=True,
    hovermode='closest',
    mapbox=go.layout.Mapbox(
        accesstoken=mapbox_access_token,
        bearing=0,
        center=go.layout.mapbox.Center(
            lat=-16,
            lon=-48
        ),
        pitch=0,
        zoom=5
    ),
)

fig.show()

import plotly.graph_objects as go

import pandas as pd
df = pd.read_csv('https://raw.githubusercontent.com/plotly/datasets/master/2011_us_ag_exports.csv')

for col in df.columns:
    df[col] = df[col].astype(str)

df['text'] = df['state'] + '<br>' + \
    'Beef ' + df['beef'] + ' Dairy ' + df['dairy'] + '<br>' + \
    'Fruits ' + df['total fruits'] + ' Veggies ' + df['total veggies'] + '<br>' + \
    'Wheat ' + df['wheat'] + ' Corn ' + df['corn']

fig = go.Figure(data=go.Choropleth(
    locations=df['code'],
    z=df['total exports'].astype(float),
    locationmode='USA-states',
    colorscale='Reds',
    autocolorscale=False,
    text=df['text'], # hover text
    marker_line_color='white', # line markers between states
    colorbar_title="Millions USD"
))

fig.update_layout(
    title_text='2011 US Agriculture Exports by State<br>(Hover for breakdown)',
    geo = dict(
        scope='usa',
        projection=go.layout.geo.Projection(type = 'albers usa'),
        showlakes=True, # lakes
        lakecolor='rgb(255, 255, 255)'),
)

fig.show()
