==> 05.1 Redefinindo DF.ipynb <==
## Redefinindo o DataFrame

import  pandas as pd
import plotly.graph_objects as go

dados = pd.read_csv('dados/novo_dados_primeira_300k_infancia.csv')
dados

dados['nome'].drop_duplicates()

qtd_cidade = dados['nome'].value_counts()
qtd_cidade

new_df = dados.drop( columns = ['Unnamed: 0', 'latitude', 'longitude','LAT', 'LOG', 'CODG_FAMILIAR', 'STAT_CADASTRO', 'DESC_STAT_CADASTRO',
       'INDI_CADASTRO_VALIDO', 'STAT_CONDICAO_CADASTRO',
       'DESC_CONDICAO_CADASTRO', 'DATA_CADASTRAM', 'DATA_ATUALZ_FAMILIA',
       'CODG_IBGE_MUN_RESIDE', 'TIPO_LOGRADOURO', 'NOME_LOGRADOURO',
       'NUMR_LOGRADOURO', 'DESC_COMPL', 'DESC_COMPL_ADICIONAL',
       'NOME_LOCALIDADE', 'NUMR_CEP_LOGRADOURO', 'INFO_REFERENCIA_LOCAL',
       'CODG_UNIDADE_TERRITAL', 'VALR_RENDA_MEDIA',
       'INDI_RECEBE_BOLSA_FAMILIA', 'INDI_BANHEIRO_DOMICILIO',
       'TIPO_DESTINO_LIXO', 'TIPO_MATERIAL_PISO',
       'TIPO_MATERIAL_PAREDE_EXTERNA', 'INDI_AGUA_CANALIZADA',
       'TIPO_CALCAMENTO', 'TIPO_ESPECIE_DOMICILIO', 'QTDE_CRIANCAS_ATE_3ANOS',
       'QTDE_CRIANCAS_ATE_6ANOS', 'QTDE_MAIOR18_NAO_SABE_LER',
       'QTDE_ENTRE6E17_NAO_FREQ_ESCOLA', 'QTDE_PESSOA_INTERNACAO',
       'VALR_DESPESA_ALUGUEL', 'INDI_FAMILIA_RECEBE_BPC',
       'INDI_RECEBE_AJUDA_DEFIC_TERC', 'DESVIO', 'aproximação']) 

#df.to_json (r'Path em que você deseja armazenar o arquivo JSON exportado \ File Name.json ')


df = new_df.to_json('dados/df_300k_sem_descritores.json',force_ascii = False)

df2 = pd.read_json('dados/df_300k_sem_descritores.json')



df = familias_por_cidade.to_csv('familias_por_cidade.csv', index = False )

mapbox_access_token = 'pk.eyJ1Ijoibml0aWR1bSIsImEiOiJjazJocXAxZnYwZ3BuM2RudjloMm9tMW51In0.2zjCD-PmPpRinCrSmlDo_Q'

def filrando_por_cidade(nome_cidade:'recebe o nome do municipo'):
    


    selecao = (new_df['nome'] == nome_cidade)
    df1 = new_df[selecao]
        
    
    site_lat = df1['lat_corrigido']
    site_lon = df1['lon_corrigido']


    fig = go.Figure(go.Scattermapbox(

            lat=site_lat,
            lon=site_lon,
            mode='markers',
            marker=go.scattermapbox.Marker(
                size=9
            ),
            text = df1['nome'],
            #text = dados['DESC_CONDICAO_CADASTRO'],


        ))

    fig.update_layout(
        autosize=True,
        hovermode='closest',
        mapbox=go.layout.Mapbox(
            accesstoken=mapbox_access_token,
            bearing=0,
            center=go.layout.mapbox.Center(
                lat=-16,
                lon=-48
            ),
            pitch=0,
            zoom=5
        ),
    )

    #fig.show()
    return fig.show()

filrando_por_cidade('Anápolis')

filrando_por_cidade('Goiânia')

filrando_por_cidade('Anhanguera')

filrando_por_cidade('Catalão')
